+++ src/snowflake/snowpark/_internal/xml_reader.py
@@ -271,26 +271,40 @@ def strip_xml_namespaces(elem: ET.Element) -> ET.Element:
 
 
 def element_to_dict_or_str(
-    element: ET.Element, attribute_prefix: str = "_", exclude_attributes: bool = False
+    element: ET.Element,
+    attribute_prefix: str = "_",
+    exclude_attributes: bool = False,
+    value_tag: str = "_VALUE",
+    null_value: str = "",
 ) -> Optional[Union[Dict[str, Any], str]]:
     """
     Recursively converts an XML Element to a dictionary.
     """
-    if not list(element) and not element.attrib:
-        return element.text.strip() if element.text and element.text.strip() else None
 
-    result: Dict[str, Any] = {}
+    def get_text(element: ET.Element) -> Optional[str]:
+        """Do not strip the text"""
+        if element.text is None or element.text == null_value:
+            return None
+        return element.text
+
+    children = list(element)
+    if not children and (not element.attrib or exclude_attributes):
+        # it's a value element with no attributes or excluded attributes, so return the text
+        return get_text(element)
+
+    result = {}
 
     if not exclude_attributes:
         for attr_name, attr_value in element.attrib.items():
-            result[f"{attribute_prefix}{attr_name}"] = attr_value
+            result[f"{attribute_prefix}{attr_name}"] = (
+                None if attr_value == null_value else attr_value
+            )
 
-    children = list(element)
     if children:
-        temp_dict: Dict[str, Any] = {}
+        temp_dict = {}
         for child in children:
             child_dict = element_to_dict_or_str(
-                child, attribute_prefix, exclude_attributes
+                child, attribute_prefix, exclude_attributes, value_tag, null_value
             )
             tag = child.tag
             if tag in temp_dict:
@@ -301,9 +315,10 @@ def element_to_dict_or_str(
                 temp_dict[tag] = child_dict
         result.update(temp_dict)
     else:
-        if element.text and element.text.strip():
-            return element.text.strip()
-
+        # it's a value element with attributes, so return the dict
+        text = get_text(element)
+        if text is not None:
+            result[value_tag] = text
     return result
 
 
@@ -318,6 +333,7 @@ def process_xml_range(
     attribute_prefix: str,
     exclude_attributes: bool,
     value_tag: str,
+    null_value: str,
     chunk_size: int = DEFAULT_CHUNK_SIZE,
 ) -> Iterator[Optional[Dict[str, Any]]]:
     """
@@ -343,6 +359,7 @@ def process_xml_range(
         attribute_prefix (str): The prefix to add to the attribute names.
         exclude_attributes (bool): Whether to exclude attributes from the XML element.
         value_tag (str): The tag name for the value column.
+        null_value (str): The value to treat as a null value.
         chunk_size (int): Size of chunks to read.
 
     Yields:
@@ -434,6 +451,8 @@ def process_xml_range(
                     element,
                     attribute_prefix=attribute_prefix,
                     exclude_attributes=exclude_attributes,
+                    value_tag=value_tag,
+                    null_value=null_value,
                 )
                 if isinstance(result, dict):
                     yield result
@@ -467,6 +486,7 @@ class XMLReader:
         attribute_prefix: str,
         exclude_attributes: bool,
         value_tag: str,
+        null_value: str,
     ):
         """
         Splits the file into byte ranges—one per worker—by starting with an even
@@ -485,6 +505,7 @@ class XMLReader:
             attribute_prefix (str): The prefix to add to the attribute names.
             exclude_attributes (bool): Whether to exclude attributes from the XML element.
             value_tag (str): The tag name for the value column.
+            null_value (str): The value to treat as a null value.
         """
         file_size = get_file_size(filename)
         approx_chunk_size = file_size // num_workers
@@ -501,5 +522,6 @@ class XMLReader:
             attribute_prefix,
             exclude_attributes,
             value_tag,
+            null_value,
         ):
             yield (element,)